+++ src/evm/config/features/fee-collector.ts
@@ -43,6 +43,7 @@ export const FEE_COLLECTOR_CHAIN_IDS = [
   EvmChainId.KATANA,
   EvmChainId.HYPEREVM,
   EvmChainId.BERACHAIN,
+  EvmChainId.PLASMA,
 ] as const
 
 type FeeCollectorChainId = (typeof FEE_COLLECTOR_CHAIN_IDS)[number]
@@ -98,6 +99,7 @@ export const UI_FEE_COLLECTOR_ADDRESS: Record<
   [EvmChainId.KATANA]: '0xde7259893af7cdbc9fd806c6ba61d22d581d5667',
   [EvmChainId.HYPEREVM]: '0xde7259893af7cdbc9fd806c6ba61d22d581d5667',
   [EvmChainId.BERACHAIN]: '0xde7259893af7cdbc9fd806c6ba61d22d581d5667',
+  [EvmChainId.PLASMA]: '0xde7259893af7cdbc9fd806c6ba61d22d581d5667',
 } as const
 export const isUIFeeCollectorChainId = isFeeCollectorChainId
 
@@ -147,6 +149,7 @@ export const SURPLUS_FEE_COLLECTOR_ADDRESS: Record<
   [EvmChainId.KATANA]: '0x5c2e112783a6854653b4bc7dc22248d3e592559c',
   [EvmChainId.HYPEREVM]: '0x5c2e112783a6854653b4bc7dc22248d3e592559c',
   [EvmChainId.BERACHAIN]: '0x5c2e112783a6854653b4bc7dc22248d3e592559c',
+  [EvmChainId.PLASMA]: '0x5c2e112783a6854653b4bc7dc22248d3e592559c',
 } as const
 export const isSurplusFeeCollectorChainId = isFeeCollectorChainId
 
@@ -196,5 +199,6 @@ export const PROTOCOL_FEE_COLLECTOR_ADDRESS: Record<
   [EvmChainId.KATANA]: '0xdbeca8fb948c42634256609bce5a3768c9b3e9eb',
   [EvmChainId.HYPEREVM]: '0xdbeca8fb948c42634256609bce5a3768c9b3e9eb',
   [EvmChainId.BERACHAIN]: '0xdbeca8fb948c42634256609bce5a3768c9b3e9eb',
+  [EvmChainId.PLASMA]: '0xdbeca8fb948c42634256609bce5a3768c9b3e9eb',
 } as const
 export const isProtocolFeeCollectorChainId = isFeeCollectorChainId