+++ src/evm/config/tokens/tokens/WETH9.ts
@@ -44,6 +44,7 @@ export const WETH9_ADDRESS = {
   [EvmChainId.KATANA]: '0xee7d8bcfb72bc1880d0cf19822eb0a2e6577ab62',
   [EvmChainId.HYPEREVM]: '0x5555555555555555555555555555555555555555',
   [EvmChainId.BERACHAIN]: '0x2f6f07cdcf3588944bf4c42ac74ff24bf56e7590',
+  [EvmChainId.PLASMA]: '0x9895d81bb462a195b4922ed7de0e3acd007c32cb',
 } as const
 
 export const WETH9 = {